Only apply strict typing to repro_zipfile
