+++ tests/test_cli.py
@@ -1,7 +1,6 @@
 from glob import glob
 import subprocess
 import sys
-from pathlib import Path
 
 from typer.testing import CliRunner
 
@@ -18,7 +17,7 @@ from tests.utils import (
 runner = CliRunner()
 
 
-def test_zip_single_file(base_path: Path) -> None:
+def test_zip_single_file(base_path):
     data_file = file_factory(base_path)
 
     rpzip_out = base_path / "rpzip.zip"
@@ -34,7 +33,7 @@ def test_zip_single_file(base_path: Path) -> None:
     assert_archive_contents_equals(rpzip_out, zip_out)
 
 
-def test_zip_directory(base_path: Path) -> None:
+def test_zip_directory(base_path):
     """Single directory, not recursive."""
     dir_tree = dir_tree_factory(base_path)
 
@@ -51,7 +50,7 @@ def test_zip_directory(base_path: Path) -> None:
     assert_archive_contents_equals(rpzip_out, zip_out)
 
 
-def test_zip_directory_recursive(base_path: Path) -> None:
+def test_zip_directory_recursive(base_path):
     """Single input directory with recursive -r flag."""
     dir_tree = dir_tree_factory(base_path)
 
@@ -68,7 +67,7 @@ def test_zip_directory_recursive(base_path: Path) -> None:
     assert_archive_contents_equals(rpzip_out, zip_out)
 
 
-def test_zip_multiple_recursive(base_path: Path) -> None:
+def test_zip_multiple_recursive(base_path):
     """Mulitiple input files with recursive -r flag."""
     dir_tree = dir_tree_factory(base_path)
 
@@ -85,7 +84,7 @@ def test_zip_multiple_recursive(base_path: Path) -> None:
     assert_archive_contents_equals(rpzip_out, zip_out)
 
 
-def test_zip_no_suffix(base_path: Path) -> None:
+def test_zip_no_suffix(base_path):
     data_file = file_factory(base_path)
 
     rpzip_out = base_path / "rpzip.zip"
@@ -101,7 +100,7 @@ def test_zip_no_suffix(base_path: Path) -> None:
     assert_archive_contents_equals(rpzip_out, zip_out)
 
 
-def test_verbosity(rel_path: Path) -> None:
+def test_verbosity(rel_path):
     """Adjustment of verbosity with -v and -q."""
     data_file = file_factory(rel_path)
     rpzip_out = rel_path / "rpzip.zip"
@@ -127,7 +126,7 @@ def test_verbosity(rel_path: Path) -> None:
     assert rpzip_result.output.strip() == ""
 
 
-def test_version() -> None:
+def test_version():
     """With --version flag."""
     result = runner.invoke(app, ["--version"])
     assert result.exit_code == 0
@@ -138,7 +137,7 @@ def test_version() -> None:
     assert output_lines[1].endswith(f"v{rpzip_version}")
 
 
-def test_python_dash_m_invocation() -> None:
+def test_python_dash_m_invocation():
     result = subprocess.run(
         [sys.executable, "-m", "rpzip", "--help"],
         capture_output=True,
